requirements.txt
bleach==6.2.0
blinker==1.9.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
Deprecated==1.2.14
Flask==3.0.3
Flask-Limiter==3.8.0
fonttools==4.54.1
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
limits==3.13.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
mysql-connector==2.2.9
mysql-connector-python==9.1.0
numpy==2.1.3
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
rich==13.9.4
seaborn==0.13.2
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.2
ua-parser==0.18.0
user-agents==2.2.0
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.16.0
requests==2.9.1
